backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.18
pillow==11.0.0
numpy==1.26.4
tensorflow==2.17.0
python-dotenv==1.0.1
pydantic==2.10.2
